refactor build setup
