--- src/Makevars.in
@@ -1,11 +1,0 @@
-PKG_CPPFLAGS=-I../inst/include @PKG_CPPFLAGS@ -DTHREADSAFE=0
-PKG_LIBS=@PKG_LIBS@
-
-.PHONY: all do_includes
-
-all: $(SHLIB)
-$(SHLIB): do_includes
-do_includes:
-	mkdir -p ../inst/include
-	cp sqlite/sqlite3.h ../inst/include
-	cp sqlite/sqlite3ext.h ../inst/include
