--- src/Makevars.win
@@ -1,18 +1,0 @@
-PKG_CPPFLAGS=-DRSQLITE_USE_BUNDLED_SQLITE \
-             -DSQLITE_ENABLE_RTREE \
-             -DSQLITE_ENABLE_FTS3 \
-             -DSQLITE_ENABLE_FTS3_PARENTHESIS \
-             -DSQLITE_SOUNDEX \
-             -DSQLITE_MAX_VARIABLE_NUMBER=40000 \
-             -DSQLITE_MAX_COLUMN=30000
-
-.PHONY: all do_includes
-
-all: $(SHLIB)
-
-$(SHLIB): do_includes
-
-do_includes:
-	mkdir -p ../inst/include
-	cp sqlite/sqlite3.h ../inst/include
-	cp sqlite/sqlite3ext.h ../inst/include
