--- src/sqlite-all.c
@@ -1,3 +1,0 @@
-#ifdef RSQLITE_USE_BUNDLED_SQLITE
-#  include "sqlite/sqlite3.c"
-#endif
