+++ src/sqlite3.h
